Extract rules
