+++ src/main/java/org/banlogic/inference/rules/InferenceRule.java
@@ -1,0 +1,10 @@
+package org.banlogic.inference.rules;
+
+import java.util.List;
+
+/**
+ * Created by Alexandru on 4/28/2016.
+ */
+public interface InferenceRule {
+    List<String> apply(String firstFormula, String secondFormula);
+}
